+++ maas-addons/maas_sale_subscription/data/ir_cron.xml
@@ -1,56 +1,28 @@
 <?xml version='1.0' encoding='utf-8'?>
-<odoo>
-  <data noupdate="0">
-
-    <!--<record id="account_analytic_cron_for_invoice" model="ir.cron">
-      <field name='name'>MyModel Sale Subscription: Generate Recurring Account Analytic</field>
-      <field name='priority'>0</field>
-      <field name='interval_number'>1</field>
-      <field name='interval_type'>months</field>
-      <field name="numbercall">-1</field>
-      <field name="doall" eval="True" />
-      <field name="model_id" ref="model_sale_subscription"/>
-      <field name="code">model.scheduler_recurring_invoice_line()</field>
-      <field name="state">code</field>
-      <field name="active">False</field>
-      <field name="nextcall" eval="(DateTime.now()+relativedelta(months=1)).strftime('%Y-%m-01 00:00:01')"/>
-    </record>
-
-    <record id="account_analytic_cron" model="ir.cron">
-      <field name='name'>MyModel Sale Subscription: Subscriptions Expiration</field>
-      <field name='priority'>2</field>
-      <field name='interval_number'>1</field>
-      <field name='interval_type'>days</field>
-      <field name="numbercall">-1</field>
-      <field name="doall" eval="True" />
-      <field name="model_id" ref="model_sale_subscription"/>
-      <field name="code">model.cron_account_analytic_account()</field>
-      <field name="state">code</field>
-      <field name="active">False</field>
-      <field name="nextcall" eval="(DateTime.now()+relativedelta(days=1)).strftime('%Y-%m-01 00:05:00')"/>
-    </record>-->
+<odoo noupdate="0">
 
     <record id="account_analytic_cron_new" model="ir.cron">
-      <field name='name'>MyModel Sale Subscription: Subscriptions Expiration and recurring invoice line</field>
-      <field name='priority'>2</field>
-      <field name='interval_number'>1</field>
-      <field name='interval_type'>days</field>
-      <field name="numbercall">-1</field>
-      <field name="doall" eval="True" />
-      <field name="model_id" ref="model_sale_order"/>
-      <field name="code">model.cron_account_analytic_account()</field>
-      <field name="state">code</field>
-      <field name="priority">0</field>
-      <field name="nextcall" eval="(DateTime.now()+relativedelta(days=1)).strftime('%Y-%m-01 20:05:00')"/>
+        <field name='name'>MyModel Sale Subscription: Subscriptions Expiration and recurring invoice
+            line</field>
+        <field name='priority'>2</field>
+        <field name='interval_number'>1</field>
+        <field name='interval_type'>days</field>
+        <field name="numbercall">-1</field>
+        <field name="doall" eval="True" />
+        <field name="model_id" ref="model_sale_order" />
+        <field name="code">model.cron_account_analytic_account()</field>
+        <field name="state">code</field>
+        <field name="priority">0</field>
+        <field name="nextcall"
+            eval="(DateTime.now()+relativedelta(days=1)).strftime('%Y-%m-01 20:05:00')" />
     </record>
 
     <record id="sale_subscription.account_analytic_cron_for_invoice" model="ir.cron">
-      <field name="active" eval="True"/>
+        <field name="active" eval="True" />
     </record>
 
     <record id="sale_subscription.account_analytic_cron" model="ir.cron">
-      <field name="active" eval="True"/>
+        <field name="active" eval="True" />
     </record>
 
-  </data>
 </odoo>
