--- maas-addons/maas_sale_subscription/data/menu.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-	<data noupdate="0">
-		<record id="menu_operation" model="website.menu">
-	    <field name="name">Operation</field>
-	    <field name="url">/operation</field>
-	    <field name="parent_id" ref="website.main_menu"/>
-	    <field name="sequence" type="int">15</field>
-	    <field name="website_id" ref="website.default_website"/>
-		</record>
-	</data>
-
-	<data noupdate="0">
-		<record id="menu_prestation_crm" model="website.menu">
-	    <field name="name">Prestation CRM</field>
-	    <field name="url">/prestation_crm</field>
-	    <field name="parent_id" ref="website.main_menu"/>
-	    <field name="sequence" type="int">15</field>
-	    <field name="website_id" ref="website.default_website"/>
-		</record>
-	</data>
-
-	<data noupdate="0">
-		<record id="menu_prestation_prm" model="website.menu">
-	    <field name="name">Prestation PRM</field>
-	    <field name="url">/prestation_prm</field>
-	    <field name="parent_id" ref="website.main_menu"/>
-	    <field name="sequence" type="int">15</field>
-	    <field name="website_id" ref="website.default_website"/>
-		</record>
-	</data>
-
-</odoo>
