+++ maas-addons/maas_sale_subscription/models/res_partner.py
@@ -19,7 +19,7 @@
 #
 ##############################################################################
 
-from odoo import api, models
+from odoo import models
 
 
 class Partner(models.Model):
