+++ maas-addons/maas_sale_subscription/views/menu.xml
@@ -1,17 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <data>
 
-        <record id="sale_subscription.menu_sale_subscription_root" model="ir.ui.menu">
-            <field name="groups_id" eval="[(4, ref('maas_base.group_subscription_management')),
-                                     (4, ref('maas_base.group_package_management'))]"/>
-        </record>
+    <record id="sale_subscription.menu_sale_subscription_root" model="ir.ui.menu">
+        <field name="groups_id"
+            eval="[(4, ref('maas_base.group_subscription_management')),
+                                     (4, ref('maas_base.group_package_management'))]" />
+    </record>
 
-        <record id="maas_sale.sale_operation_menu_root" model="ir.ui.menu">
-            <field name="parent_id" ref="sale_subscription.menu_sale_subscription_root"/>
-            <field name="sequence">5</field>
-        </record>
-
-
-    </data>
+    <record id="maas_sale.sale_operation_menu_root" model="ir.ui.menu">
+        <field name="parent_id" ref="sale_subscription.menu_sale_subscription_root" />
+        <field name="sequence">5</field>
+    </record>
 </odoo>
