+++ maas-addons/maas_sale_subscription/views/sale_order_view.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="sale_subscription_view_form_inherit" model="ir.ui.view">
+        <field name="name">sale.subscription.view.form.inherit</field>
+        <field name="model">sale.order</field>
+        <field name="priority">999</field>
+        <field name="inherit_id" ref="sale_subscription.sale_subscription_order_view_form" />
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="next_invoice_date"
+                    attrs="{'invisible': ['|', ('stage_category', 'not in', ['progress', 'paused']) , ('recurrence_id', '=', False)],
+                                   'required': [('recurrence_id', '!=', False), ('state', 'in', ['sale', 'done'])]}"
+                    position="move" />
+                <field name="service_start_date"
+                    attrs="{'invisible': [('recurrence_id', '=', False)]}" />
+                <field name="balance" attrs="{'invisible': [('recurrence_id', '=', False)]}" />
+            </field>
+            <field name="date_order" position="after">
+                <field name="allowance" attrs="{'invisible': [('recurrence_id', '=', False)]}" />
+                <field name="start_date" position="move"
+                    attrs="{'invisible': [('recurrence_id', '=', False)]}" />
+            </field>
+            <field name="payment_term_id" position="before">
+                <field name="package_id" options="{'no_create': False, 'no_open': False}"
+                    string="Basic Package"
+                    attrs="{'invisible': [('recurrence_id', '=', False)]}" />
+                <field name="current_package_id"
+                    options="{'no_create': True, 'no_open': True}"
+                    readonly="1"
+                    force_save="1"
+                    string="Current Package" attrs="{'invisible': [('recurrence_id', '=', False)]}" />
+            </field>
+            <xpath expr="//field[@name='order_line']/tree/field[@name='product_id']"
+                position="before">
+                <field name="date"
+                    attrs="{'column_invisible': [('parent.recurrence_id', '=', False)]}" />
+                <field name="period"
+                    attrs="{'column_invisible': [('parent.recurrence_id', '=', False)]}" />
+            </xpath>
+            <xpath expr="//field[@name='order_line']/tree/field[@name='discount']" position="after">
+                <field name="qty_consumed"
+                    attrs="{'column_invisible': [('parent.recurrence_id', '=', False)]}" />
+                <field name="qty_cumulative"
+                    attrs="{'column_invisible': [('parent.recurrence_id', '=', False)]}" />
+            </xpath>
+            <xpath expr="//field[@name='order_line']/tree/field[@name='price_subtotal']"
+                position="after">
+                <field name="state_subscription"
+                    attrs="{'column_invisible': [('parent.recurrence_id', '=', False)]}" />
+            </xpath>
+        </field>
+    </record>
+
+    <!-- Sale Subscription Tree View -->
+    <record id="sale_subscription_view_tree_inherit" model="ir.ui.view">
+        <field name="name">sale.subscription.view.tree.inherit</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale_subscription.sale_subscription_view_tree" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='next_invoice_date']" position="attributes">
+                <attribute name="invisible">1</attribute>
+            </xpath>
+            <xpath expr="//field[@name='amount_total']" position="attributes">
+                <attribute name="invisible">1</attribute>
+            </xpath>
+            <xpath expr="//field[@name='partner_id']" position="after">
+                <field name="start_date" />
+                <field name="end_date" />
+            </xpath>
+        </field>
+    </record>
+
+    <!-- Sale Subscription Search View &ndash;&gt; -->
+    <record id="sale_subscription_view_search_inherit" model="ir.ui.view">
+        <field name="name">sale.subscription.search.inherit</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale_subscription.sale_subscription_view_search" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[last()]" position="after">
+                <field name="start_date" />
+                <field name="service_start_date" />
+                <field name="package_id" />
+                <field name="state" />
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
